Revert "Feat/34 usercontrolveiw UI"
